requirements.txt
imageio==2.3.0
numpy==1.13.3
matplotlib==2.1.0
scipy==0.19.1
tqdm==4.19.5
opencv_python==3.4.0.12
psutil==5.4.0
Pillow==5.3.0
tifffile==0.15.0
